+++ uui/components/widgets/Avatar.tsx
@@ -1,4 +1,15 @@
-import { withMods } from '@epam/uui-core';
+import { devLogger, withMods } from '@epam/uui-core';
 import { Avatar as uuiAvatar, AvatarProps } from '@epam/uui-components';
 
-export const Avatar = withMods<AvatarProps>(uuiAvatar, () => []);
+export const Avatar = withMods<AvatarProps>(
+    uuiAvatar,
+    () => [],
+    (props) => {
+        if (__DEV__) {
+            if (props.onClick) {
+                devLogger.warn('Avatar: Property onClick is deprecated and will be removed in the future release.');
+            }
+        }
+        return null;
+    },
+);